feat : add smooth scroll to Skills section from Intro
Introduced a scrollToTarget() method in IntroComponent to smoothly scroll to the Skills section when the 'Explore the rest' button is clicked. Updated the skills section to use an 'id' for targeting and refactored related class names for consistency. Also commented out the header in main.component.html (will not implement for now).